+++ ml/src/moomoo_ml/db.py
@@ -3,10 +3,12 @@
 import datetime
 import os
 import re
+from pathlib import Path
 from typing import Annotated, ClassVar
 
+import numpy as np
 from pgvector.sqlalchemy import Vector
-from sqlalchemy import create_engine, func
+from sqlalchemy import create_engine, func, text
 from sqlalchemy.dialects import postgresql
 from sqlalchemy.orm import DeclarativeBase, Mapped, Session, mapped_column
 
@@ -41,6 +43,38 @@ class FileEmbedding(BaseTable):
     insert_ts_utc: Mapped[datetime.datetime] = mapped_column(
         nullable=False, server_default=func.current_timestamp(), index=True
     )
+
+    @classmethod
+    def fetch_numpy_embeddings(
+        cls, only_unconditioned: bool = False
+    ) -> tuple[list[Path], np.ndarray]:
+        """Get embeddings from the database.
+
+        Returns a list of Paths and a 2d numpy array of embeddings. Each row in the array
+        corresponds to the embedding of the file at the same index in the list.
+
+        If only_unconditioned is provided, only embeddings which have not been conditioned will be
+        returned.
+        """
+        if only_unconditioned:
+            conditioner_sql = cls.conditioned_embedding.is_(None)
+        else:
+            conditioner_sql = text("true")
+
+        with get_session() as session:
+            query = (
+                session.query(cls)
+                .filter(cls.success.is_(True))
+                .filter(conditioner_sql)
+                .order_by(cls.filepath)
+            )
+
+            if not query.count():
+                paths, embeddings = [], []
+            else:
+                paths, embeddings = zip(*[(Path(i.filepath), i.embedding) for i in query.all()])
+
+        return paths, np.array(embeddings)
 
 
 class LocalFileExcludeRegex(BaseTable):
